Se corrigen errores en ManejadorDeArchivos a la hora de generar las promociones

+++ src/parque/PromocionPorcentual.java
@@ -3,7 +3,7 @@ package parque;
 public class PromocionPorcentual extends Promocion {
     private double porcentajeDescuento;
 
-    public PromocionPorcentual(String nombre, int porcentajeDescuento) {
+    public PromocionPorcentual(String nombre, double porcentajeDescuento) {
     	super(nombre);
         this.porcentajeDescuento = porcentajeDescuento;
     }
